Fix backend, prepare to development deploy

+++ backendGo/Driver/dbSet.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 )
 
-func GetSet(routineID, exerciseID string) (structmodels.Set, error) {
+func GetSet(routineID, exerciseID string, userID int) (structmodels.Set, error) {
 	// Uso LIMIT 1 para devolver unicamente un valor, despues habria que hacer un order
 	// habra que asignar un usuario o verificar de que rutina se pide (ya que va asociado a un user)
 
@@ -13,7 +13,7 @@ func GetSet(routineID, exerciseID string) (structmodels.Set, error) {
 	//query_user := "SELECT id FROM Routines WHERE user_id = ?"
 	//err := db.QueryRow(query_user, routineID).Scan(&Set.RoutineID)
 
-	query := "SELECT id, routine_id, exercise_id, set_number, reps, weight FROM Sets WHERE routine_id = ? AND exercise_id = ? LIMIT 1"
+	query := fmt.Sprintf("SELECT  routine_id, exercise_id, set_number, reps, weight FROM Sets WHERE routine_id = ? AND exercise_id = ? AND user_id = %d LIMIT 1", userID)
 	err := db.QueryRow(query, routineID, exerciseID).Scan(&Set.Set_number, &Set.Reps, &Set.Weight, &Set.RoutineID, &Set.ExerciseID)
 
 	if err != nil {
@@ -23,10 +23,10 @@ func GetSet(routineID, exerciseID string) (structmodels.Set, error) {
 	return Set, nil
 }
 
-func GetAlSet(routineID, exerciseID string) ([]structmodels.Set, error) {
+func GetAlSet(routineID, exerciseID string, userID int) ([]structmodels.Set, error) {
 
 	var Sets []structmodels.Set
-	query := fmt.Sprintf("SELECT id, routine_id, exercise_id, set_number, reps, weight FROM Sets WHERE routine_id = %s AND exercise_id = %s;", routineID, exerciseID)
+	query := fmt.Sprintf("SELECT  routine_id, exercise_id, set_number, reps, weight FROM Sets WHERE routine_id = %s AND exercise_id = %s AND user_id = %d ;", routineID, exerciseID, userID)
 	rows, err := db.Query(query)
 
 	if err != nil {
@@ -36,7 +36,7 @@ func GetAlSet(routineID, exerciseID string) ([]structmodels.Set, error) {
 
 	for rows.Next() {
 		var Set structmodels.Set
-		if err := rows.Scan(&Set.ID, &Set.RoutineID, &Set.ExerciseID, &Set.Set_number, &Set.Reps, &Set.Weight); err != nil {
+		if err := rows.Scan(&Set.RoutineID, &Set.ExerciseID, &Set.Set_number, &Set.Reps, &Set.Weight); err != nil {
 			return nil, err
 		}
 		Sets = append(Sets, Set)
@@ -45,10 +45,10 @@ func GetAlSet(routineID, exerciseID string) ([]structmodels.Set, error) {
 	return Sets, nil
 }
 
-func GetAlSetRoutine(routineID string) ([]structmodels.Set, error) {
+func GetAlSetRoutine(routineID string, userID int) ([]structmodels.Set, error) {
 
 	var Sets []structmodels.Set
-	query := fmt.Sprintf("SELECT id, routine_id, exercise_id, set_number, reps, weight FROM Sets WHERE routine_id = %s;", routineID)
+	query := fmt.Sprintf("SELECT  routine_id, exercise_id, set_number, reps, weight FROM Sets WHERE routine_id = %s; AND user_id = %d", routineID, userID)
 	rows, err := db.Query(query)
 
 	if err != nil {
@@ -58,7 +58,7 @@ func GetAlSetRoutine(routineID string) ([]structmodels.Set, error) {
 
 	for rows.Next() {
 		var Set structmodels.Set
-		if err := rows.Scan(&Set.ID, &Set.RoutineID, &Set.ExerciseID, &Set.Set_number, &Set.Reps, &Set.Weight); err != nil {
+		if err := rows.Scan(&Set.RoutineID, &Set.ExerciseID, &Set.Set_number, &Set.Reps, &Set.Weight); err != nil {
 			return nil, err
 		}
 		Sets = append(Sets, Set)
@@ -67,9 +67,9 @@ func GetAlSetRoutine(routineID string) ([]structmodels.Set, error) {
 	return Sets, nil
 }
 
-func PutSet(RoutineID, ExerciseID, Set_number, Weight, Reps string, userID int) error {
-	query := `UPDATE Sets SET reps = ?, weight = ? WHERE routine_id = ? AND exercise_id = ? AND set_number = ?`
-	_, err := db.Exec(query, Reps, Weight, RoutineID, ExerciseID, Set_number)
+func PutSet(Set structmodels.Set, userID int) error {
+	query := fmt.Sprintf(`UPDATE Sets SET reps = ?, weight = ? WHERE routine_id = ? AND exercise_id = ? AND user_id= %d AND set_number = ? AND id = ?`, userID)
+	_, err := db.Exec(query, Set.Reps, Set.Weight, Set.RoutineID, Set.ExerciseID, Set.Set_number)
 
 	if err != nil {
 		fmt.Println("Error updating set:", err)
@@ -78,10 +78,28 @@ func PutSet(RoutineID, ExerciseID, Set_number, Weight, Reps string, userID int) 
 	return nil
 }
 
+func PostSet(Set structmodels.Set, userID int) error {
+	query := `
+        INSERT INTO Sets (routine_id, exercise_id, set_number, weight, reps, user_id) 
+        VALUES (?, ?, ?, ?, ?, ?)
+    `
+	_, err := db.Exec(query, Set.RoutineID, Set.ExerciseID, Set.Set_number, Set.Weight, Set.Reps, userID)
+	if err != nil {
+		fmt.Println("Error inserting set:", err)
+		return err
+	}
+	return nil
+}
+
 func DelSet(RoutineID, ExerciseID, Set_number string, userID int) error {
 	var setID int
-	queryGetID := "SELECT id FROM Sets WHERE routine_id = ? AND exercise_id = ? AND set_number = ?"
-	db.QueryRow(queryGetID, RoutineID, ExerciseID, Set_number).Scan(&setID)
+	var dbUserID int
+	queryGetID := "SELECT id , user_id FROM Sets WHERE routine_id = ? AND exercise_id = ? AND set_number = ?"
+	db.QueryRow(queryGetID, RoutineID, ExerciseID, Set_number).Scan(&setID, &dbUserID)
+
+	if dbUserID != userID {
+		return fmt.Errorf("userID does not match the owner of the set")
+	}
 
 	query := "DELETE FROM Sets WHERE id = ?;"
 	_, err := db.Exec(query, setID)
